+++ R/tab_rate.R
@@ -18,10 +18,10 @@
 #' @examples
 #' set_survey(namcs2019sv)
 #' # pop is a data frame
-#' tab_rate("MSA", uspop2019$MSA)
+#' tab_rate("MSA", uspop_example$MSA)
 #'
 #' # pop is a single number
-#' tab_rate("MDDO", uspop2019$total)
+#' tab_rate("MDDO", uspop_example$total)
 tab_rate = function(vr, pop
   , per = getOption("surveytable.rate_per")
   , drop_na = getOption("surveytable.drop_na")