+++ R/tab_subset_rate.R
@@ -19,7 +19,7 @@
 #'
 #' @examples
 #' set_survey(namcs2019sv)
-#' tab_subset_rate("AGER", "SEX", uspop2019$`AGER x SEX`)
+#' tab_subset_rate("AGER", "SEX", uspop_example$`AGER x SEX`)
 tab_subset_rate = function(vr, vrby
                            , pop
                            , lvls = c()